Add suffix 'Controller' to file name and rename entry method to 'Index'
